Static Templates made
Landlord Home, Property Page and Add Property were made. I also made a
template.html for the general in app styling. A lot of the stuff on the
pages is going to be server generated so I put as much I thought was
good.

Also I don't know anything about forms and Django/Python so I haven't
done those yet.